FileTransfer now opens and writes files to defined sync directory.

+++ SyncCFT/Client.cc
@@ -257,7 +257,7 @@ bool Client::completeFileTransfer(Message* msg, bool first) {
     int tries = 0;
     while(!ready){
         
-        if(!first && !_trns->recv(msg, timeout)) {
+        if(!first && !_trns->recv(msg, (int)timeout)) {
             cout << "[CLIENT] Wait FILE timeout" << endl;
             if (tries++ > CLIENT_CHUNK_RETRIES) {
                 cout << "[CLIENT] Too many retries" << endl;
